Changing RRTMGP repo link
From: https://github.com/earth-system-radiation/rte-rrtmgp.git
Tag: v1.7
To: https://github.com/EarthWorksOrg/rte-rrtmgp.git
Tag: v1.7_ew_release_2.2
This is a temporary measure for the functional release of Earthworks.
This change fixes the GPU seg fault in RRTMGP because of the missing
data directives in the CAM-RRTMGP interface. This change should be
removed once we add the data directives to the CAM-RRTMGP interface.