Workaround issues with MPASv8 ncdata files
This commit comments out the real-data initial condition files and
corresponding topography files that were added in EWOrg/CAM PR#16. These
files cause the chemistry solvers to fail to converge, the dynamics to
become unstable (NaNs in the w field), and the runs of these cases to
fail.

By turning off gravity wave frontogenesis the other, existing MPAS
real-data files can be used with the cam_dev physics suite. This commit
also modifies logic in build-namelist so that frontogenesis is only
turned on with cam_dev if the MPAS-dycore isn't being used.